feat: Add RSA KeyStore attestation support
Currently, the application only allows for verifying ECDSA keys as it lacked support for attesting RSA keys within the Android KeyStore.
With this commit, it allows to perform a complete validation of the two currently supported asymmetric key types in the Android KeyStore.
A new menu option "Attest RSA key" was added for this purpose.
Additionally, when exporting the keychain in p7b format, the keystore's type will be included as part of the filename to differentiate between the ECDSA (current default) and RSA.

Signed-off-by: linuxct <linuxct@linuxct.space>